requirements.txt
# SPIDEY TODO — Python dependencies
# Install with:  pip install -r requirements.txt

# ── File reading ──────────────────────────
pypdf==4.3.1            # Read PDF files line by line
python-docx==1.1.2      # Read .docx Word documents

# ── Browser injection (optional) ─────────
websocket-client==1.8.0 # Push tasks into Chrome via CDP (--inject flag)

# ── Web server for hosting ────────────────
flask==3.0.3            # Lightweight web server
flask-cors==4.0.1       # Allow browser requests from any origin
gunicorn==22.0.0        # Production WSGI server (Linux/Mac)
waitress==3.0.0         # Production WSGI server (Windows)
